requirements.txt
scylla-driver==3.25.*
psycopg[binary]==3.2.*
kafka-python==2.0.2